+++ src/org/ods/orchestration/FinalizeStage.groovy
@@ -242,9 +242,10 @@ class FinalizeStage extends Stage {
         )
 
         // export the cd namespace into the RM repo as template - to be able to recover it.
-
         new FinalizeOdsComponent(project, steps, git, logger).
             run(['id': "${FinalizeOdsComponent.RELEASE_MANAGER_REPO_ID}"], '.', false)
+
+        // open - do we want to merge the state of -cd back into master? ..
 
         if (project.isWorkInProgress) {
             git.pushRef('master')