+++ src/org/ods/services/OpenShiftService.groovy
@@ -20,6 +20,7 @@ class OpenShiftService {
     static final String ROLLOUT_WAITING = 'waiting'
     static final String DEPLOYMENTCONFIG_KIND = 'DeploymentConfig'
     static final String DEPLOYMENT_KIND = 'Deployment'
+    static final String TAILOR_FILE_NAME = 'Tailorfile'
 
     private final IPipelineSteps steps
     private final ILogger logger